# core-banking-cards-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-banking-cards
  description: Banking Cards Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/cards
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/bins/{binId}:
    get:
      tags:
        - BINs
      summary: Get BIN by ID
      description: |-
        Retrieve a specific Bank Identification Number (BIN) record by its unique identifier.

        ## Description
        This endpoint returns detailed information about a BIN, including its associated issuer, card network, and card type. This information is essential for card issuance and transaction processing.

        ## Response Details
        The response includes comprehensive information about the requested BIN:
        * BIN number and length
        * Associated issuer details
        * Card network information
        * Card type specifications
        * Geographic and currency information
        * Activation status and description

        ## Path Parameters
        * `binId` - The unique identifier of the BIN record in the system

        ## Common Use Cases
        * Retrieving BIN details before card issuance
        * Verifying BIN configuration during system setup
        * Checking BIN status during transaction processing
      operationId: getBIN
      parameters:
        - name: binId
          in: path
          description: Unique identifier of the BIN
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the BIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BINDTO'
        '404':
          description: BIN not found
    put:
      tags:
        - BINs
      summary: Update BIN
      description: |-
        Update an existing Bank Identification Number (BIN) record by its unique identifier.

        ## Description
        This endpoint allows modification of BIN attributes such as associated card type, network, country code, currency code, and activation status.

        ## Updatable Fields
        * `cardTypeId` - Type of card issued under this BIN
        * `cardNetworkId` - Card network associated with this BIN
        * `countryCode` - ISO country code for the BIN
        * `currencyCode` - ISO currency code for cards issued under this BIN
        * `isActive` - Whether the BIN is active
        * `description` - Additional information about the BIN

        ## Important Note
        ⚠️ The BIN number itself should typically not be changed as it is a standardized identifier assigned by card networks. Changing the BIN number could cause transaction routing issues and affect existing cards issued with this BIN.

        ## Path Parameters
        * `binId` - The unique identifier of the BIN record to update

        ## Common Use Cases
        * Updating BIN status (activating or deactivating)
        * Modifying currency or country information
        * Updating card type or network associations
      operationId: updateBIN
      parameters:
        - name: binId
          in: path
          description: Unique identifier of the BIN to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BINDTO'
        required: true
      responses:
        '200':
          description: BIN updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BINDTO'
        '400':
          description: Invalid BIN data provided
        '404':
          description: BIN not found
    delete:
      tags:
        - BINs
      summary: Delete BIN
      description: |-
        Delete a Bank Identification Number (BIN) record by its unique identifier.

        ## Description
        This endpoint permanently removes a BIN from the system. It completely erases the BIN record from the database.

        ## Warning
        ⚠️ **This operation should be used with extreme caution!**

        Deleting a BIN has several important implications:
        * Cards issued with this BIN may experience transaction routing issues
        * Card issuance processes using this BIN will fail
        * Historical transaction data may become inconsistent

        ## Recommended Alternative
        In most production scenarios, BINs should be **deactivated** rather than deleted if they are no longer in use. This preserves the BIN record for historical purposes while preventing new cards from being issued with this BIN.

        ## Appropriate Use Cases
        BIN deletion should be limited to:
        * Test BINs in non-production environments
        * BINs created in error that have never been used for card issuance
        * BINs that are no longer assigned to your organization by the card network

        ## Path Parameters
        * `binId` - The unique identifier of the BIN record to delete

        ## Required Permissions
        This operation typically requires elevated administrative privileges.
      operationId: deleteBIN
      parameters:
        - name: binId
          in: path
          description: Unique identifier of the BIN to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: BIN deleted successfully
        '404':
          description: BIN not found
...
  /api/v1/bins:
    get:
      tags:
        - BINs
      summary: Get BIN by Number
      description: |-
        Retrieve a specific Bank Identification Number (BIN) record by its BIN number.

        ## Description
        This endpoint allows looking up BIN information using the actual BIN number (first 6-8 digits of a card), which is useful during transaction processing and card validation workflows.

        ## Query Parameters
        * `number` - The BIN number to search for (e.g., '411111' for a Visa card)

        ## Response Details
        The response includes the same comprehensive BIN information as the 'Get BIN by ID' endpoint, but the lookup is performed using the actual BIN number rather than the internal database identifier.

        ## Common Use Cases
        * Identifying the issuer of a card during transaction processing
        * Determining card type and network during payment validation
        * Verifying card origin during fraud detection processes
        * Routing transactions based on BIN information
      operationId: getAllBINs_1
      parameters:
        - name: paginationRequest
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/PaginationRequest'
        - name: number
          in: query
          description: BIN number
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved the BIN
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/BINDTO'
                  - $ref: '#/components/schemas/PaginationResponse'
        '404':
          description: BIN not found
    post:
      tags:
        - BINs
      summary: Create BIN
      description: |-
        Create a new Bank Identification Number (BIN) record.

        ## Description
        This endpoint allows financial institutions to register new BINs in the system. The BIN is associated with a specific issuer, card network, and card type, and is used to identify the origin of payment cards.

        ## Required Fields
        * `binNumber` - The BIN digits (typically 6-8 digits)
        * `binLength` - Length of the BIN (6, 8, etc.)
        * `issuerId` - Financial institution that owns this BIN
        * `cardNetworkId` - Card network associated with this BIN
        * `cardTypeId` - Type of card issued under this BIN

        ## Optional Fields
        * `countryCode` - ISO country code for the BIN
        * `currencyCode` - ISO currency code for cards issued under this BIN
        * `isActive` - Whether the BIN is active
        * `description` - Additional information about the BIN

        ## Regulatory Considerations
        BINs are typically assigned by card networks to financial institutions. Before creating a BIN, ensure that your organization has been officially assigned this BIN by the relevant card network. Using unauthorized BINs can lead to transaction routing issues and regulatory violations.
      operationId: createBIN
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BINDTO'
        required: true
      responses:
        '201':
          description: BIN created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BINDTO'
        '400':
          description: Invalid BIN data provided
...
components:
  schemas:
...
    BINDTO:
      required:
        - binLength
        - binNumber
        - cardNetworkId
        - cardTypeId
        - countryCode
        - currencyCode
        - isActive
        - issuerId
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        binId:
          type: string
          format: uuid
          readOnly: true
        binNumber:
          type: string
        binLength:
          maximum: 8
          minimum: 6
          type: integer
          format: int32
        issuerId:
          type: string
          format: uuid
        cardNetworkId:
          type: string
          format: uuid
        cardTypeId:
          type: string
          format: uuid
        countryCode:
          pattern: ^[A-Z]{2}$
          type: string
        currencyCode:
          pattern: ^[A-Z]{3}$
          type: string
        isActive:
          type: boolean
        description:
          maxLength: 500
          minLength: 0
          type: string
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...